Remove DB Migration Runner
DB Migrations (flyway) are run out-of-band.  They require you to
first `sbt assembly` and build a fat jar, and then subsequently
to run a `DbMigrationRunner`.

All of those steps are unnecessary as flyway supports concurrent
attempts to run migrations.  The specific item is in the FAQ
at https://flywaydb.org/documentation/faq#parallel

Instead, just run migrations when the application starts.

* Modified `VinylDNSJDBC` to _always_ run database migrations.  Removed
the embedded flag as that is no longer needed.
* Deleted the `migrations.conf` file as that will no longer be used.
* Cleaned up the config files.  The salient change was to remove the
`local-mode` flag, as it is no longer needed.  Found some left over
cruft along the way that is unused and cleared that up.