Log for all monitors (#94)
The amount of VinylDNS traffic is generally rather low.  It makes
sense to log every monitored event in order to calculate error
rate, latency, and volume.  All of those are easy to generate with
this change from standard log tools like Kibana and Splunk.

* Added a `logEntry` function to `Monitor` that creates a log entry
as a string
* Updated the `Monitor` to log to info for every capture or failure

In addition, removed the `log4jReporter`.  Logging the metrics in
bulk added a lot of noise to the logs.  Plus, cannot calculate
histograms across machines, which can be done by simply logging
all events.

* Removed log4jReporter
* Removed conf entries related to log4j reporter
* Cleaned up some more config that is no longer used