Update to sbt-protoc (#113)
The sbt protobuf plugin we were using forced developers to install
protoc version 2.6.1 on their local.  That makes it difficult
to onboard new developers.

sbt-protoc is an alternative plugin.  It has a lot of features we
are presently not using.  The biggest feature it brings is to
not require developers to install protoc.

* build.sbt - use the new protoc plugin
* VinylDNSProto.proto - add syntax = proto2 to ensure compile
to compatible protobuf version
* plugins.sbt - remove the old protobuf plugin
* protoc.sbt - add the new protoc plugin
* Removing ProtocPlugin as it is an auto-plugin
* Remove uneeded protoc lib dependency
* Update docs to remove protoc requirement
* Remove protobuf install from travis