Adding publishing to sonatype
Tested deploying core to snapshots and it worked

* Updated plugins.sbt to include the sonatype plugin as well as
the pgp plugin.  The pgp plugin is new, and simply pulls gpg keys
from your default keyring
* Updated build.sbt to publish to sonatype.

You need to have a file ~/.sbt/1.0/sonatype.sbt that has credentials
to publish to sonatype.

You also need to have a file ~/.ivy2/.gpgCredentials that has
the gpg credentials used to load the key to sign artifacts with.