Fix authentication error in the portal
The root cause for the authentication error is that the portal
was not decrypting the user secret key before signing requests.

This is solved via the following:

1. Update the `UserAccountAccessor`, make sure that whenever a
user is returned, the secret key is decrypted.
1. Make sure that the `encrypt-user-secrets` feature flag is `on`
in the API reference.conf.  This was why in testing locally we
did not hit the same issue that we saw in the development environment.
Because the flag was false, test users secrets were not encrypted.

* `portal application.conf` - set the crypto to match the API
* `Dependencies.scala` - eliminate some duplication of dependencies
* `api reference.conf` - set the encrypt-user-secrets flag to true
* `TestApplicationData.scala` - modify the mock play app to have a
CryptoAlgebra binding
* `UserAccountAccessor` - decrypt the user secret key when retrieving
the user
* `VinylDNSModule` - add binding for CryptoAlgebra for dependency
injection.