Fix docker releases
There was an issue starting the docker containers due to how native packager
works where we were seeing issues with the container being able to start.

The issue was that we were assuming a "daemon" user to run the containers under.
At some point this changed to "1001:0".  As a result, there were not sufficient
privileges to start the containers because the "daemon" user was invalid or did
not have access to the scripts created by sbt native packager.

* `build.sbt` - update the user to "1001:0" for our custom install.  Cleaned up the hardcoded references in the script extras to `/opt/docker` to use the variable `app_home` instead.
* `plugins.sbt` - updated to the latest sbt native packager