Support non ActiveDirectory LDAP (#859)
Support non ActiveDirectory LDAP

This PR has an optional local portal setup against this docker container - https://github.com/rroemhild/docker-test-openldap

The base modifications for LDAP was to change the actual authentication flow.  Before, we only attempted to bind (setting up a DirContext and relying on an exception).  We would test all of the search bases until we exhausted the list.

The new approach works differently:

1. First, login using the main service account
2. Second, do a lookup of the user
3. Finally, attempt to bind to that user's context directly using the password provided.

This works fine with both AD LDAP as well as the example docker container which uses OpenLDAP

Besides these changes, need to make configurable the userNameField, which is the ldap attribute that is used to search for the username sent in the login screen.  In AD, this is `sAMAccountName`, but in the example it is `uid`, the logon field is up to the way LDAP is setup

- `docker-up-vinyldns.sh` - fixed a quote issue with the startup script to properly send in the version of vinyldns
- `docker-compose-build.yml` - added the `ldap` container so the portal can connect as `vinyldns-ldap`
- `docker/portal/application.conf` - new config file so that we can connect to the new ldap container
- `docker-compose.yml` - added the `ldap` container here as well so we can play with it using `reStart` in sbt instead of `docker-up-vinyldns.sh` - simplifies local testing
- `LdapAuthenticator.scala` - this is where the main changes happen