.asHealthCheck takes a class param for reporting (#954)
* .asHealthCheck takes a class param for reporting

Previously if a health check failed, there was no way of knowing _which_
one failed, unless you can isolate the particular type of failure --
though they're often very general (connection timeout, etc). If a health
check fails, it's important to know which one, or at least to log which
one it was.

I'm not sure if there's a way to infer the calling class without doing
some gnarly and expensive call-stack stuff, so here's a simple fix. If
there's a better way, I'm all ears.

I did experiment a little bit parameterizing the HealthCheck, but that
honestly seemed like more trouble than it was worth.