Make tenant id optional (#1031)
Tenant id is not uniformly used across all OIDC providers (such as keycloak).

Make tenant id in configuration and the corresponding OIDC flow check optional for the time being.  

Need to overhaul the OIDC portal code which is well underway but not ready yet, so this is a temporary workaround.