Remove duplicate log output
Appenders in logback are additive [0], specifying them multiple times
leads to log lines being output multiple times. It is usually enough to
set the appender on the root logger, and nowhere else.
Setting the same log level on multiple package levels is also kind of
unneeded, the top-most one would be enough; but not touching that here.

[0] https://logback.qos.ch/manual/configuration.html#cumulative